+++ storage/src/testFixtures/java/tech/pegasys/teku/storage/store/StoreAssertions.java
@@ -54,6 +54,7 @@ public class StoreAssertions {
             "stateProvider",
             "checkpointStates",
             "forkChoiceStrategy",
+            "maybeEpochStates",
             "epochStatesCountGauge")
         .isEqualTo(expectedState);
     assertThat(actualState.getOrderedBlockRoots())